+++ source/libs/distrho-plugin-toolkit/DistrhoUtils.h
@@ -102,7 +102,7 @@ public:
 
     d_string(int value)
     {
-        size_t strBufSize = abs(value/10) + 3;
+        size_t strBufSize = (unsigned int)abs(value/10) + 3;
         char   strBuf[strBufSize];
         snprintf(strBuf, strBufSize, "%d", value);
